substring

+++ Day7/src/test/java/testBase/BaseClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Date;
@@ -13,23 +14,27 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
+import org.openqa.selenium.remote.RemoteWebDriver;
 
 public class BaseClass {
 	public static WebDriver driver;
 	
 	public Logger logger;
 	public Properties p;
-	 @BeforeClass(groups= {"Sanity", "Regression", "Master" })
+	 @SuppressWarnings("deprecation")
+	@BeforeClass(groups= {"Sanity", "Regression", "Master" })
 	 
 	 @Parameters ({"os", "browser"})
 	 public	void setup( String os ,String br) throws IOException {
@@ -42,14 +47,49 @@ public class BaseClass {
 		 
 		 logger = LogManager.getLogger(this.getClass());
 		 
-		 switch(br.toLowerCase()) {
-		 case "chrome" :driver = new ChromeDriver(); break;
-		 case "edge" :driver = new EdgeDriver(); break;
-		 case "firefox" :driver = new FirefoxDriver(); break;
-
-		  default: System.out.println("Invalid browser name..."); return;
-		 }
+		 if(p.getProperty("execuation_env").equalsIgnoreCase("remote"))
+		 {
+			 DesiredCapabilities capabilities = new DesiredCapabilities();
+			// capabilities.setPlatform(Platform.WIN11);
+		//	 capabilities.setBrowserName("chrome");
+		 // os
+			 if(os.equalsIgnoreCase("windows")) {
+				 capabilities.setPlatform(Platform.WIN11);
+			 } else {
+				 System.out.println("No matching os");
+				 
+				 return;
+			 }
+		 // browser
+			 switch(br.toLowerCase()) 
+			 {
+			 case "chrome": 	 capabilities.setBrowserName("chrome"); break;
+			 case "edge" : capabilities.setBrowserName("MicrosoftEdge"); break;
+			 default: System.out.println("No ,matching browser"); 
+			 return;
+			 }
+		 driver = new RemoteWebDriver(new URL("http://localhost:4444/wd/hub" ), capabilities);
 		 
+		 } 
+		 
+		 
+		 // this is for local
+		 
+		 if(p.getProperty("execuation_env").equalsIgnoreCase("local"))
+          {
+			 switch(br.toLowerCase()) 
+			 {
+			 case "chrome" :driver = new ChromeDriver(); break;
+			 case "edge" :driver = new EdgeDriver(); break;
+			 case "firefox" :driver = new FirefoxDriver(); break;
+
+			  default: System.out.println("Invalid browser name..."); return;
+			 }
+			  
+		 }
+	 
+		 
+		
 		 
 		//	driver = new ChromeDriver();
 			driver.manage().deleteAllCookies();
